# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/users/me/password:
    put:
      summary: 'Update Password'
      operationId: updatePassword
      description: "Update the authenticated user's password."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Password updated successfully'
                properties:
                  message:
                    type: string
                    example: 'Password updated successfully'
        400:
          description: ''
          content:
            application/json:
              schema:
                oneOf:
                  -
                    description: ''
                    type: object
                    example:
                      message: 'Current password is incorrect'
                    properties:
                      message:
                        type: string
                        example: 'Current password is incorrect'
                  -
                    description: ''
                    type: object
                    example:
                      message: 'Password was used before'
                    properties:
                      message:
                        type: string
                        example: 'Password was used before'
      tags:
        - 'User Management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: 'The current password.'
                  example: CurrentPass123!
                  nullable: false
                password:
                  type: string
                  description: 'The new password.'
                  example: NewStrongPass123!
                  nullable: false
                password_confirmation:
                  type: string
                  description: 'The new password confirmation.'
                  example: NewStrongPass123!
                  nullable: false
              required:
                - current_password
                - password
                - password_confirmation
...